[Backend] Refactored the Account Forces endpoint
