+++ Backend/src/api/accounts/controllers/index.ts
@@ -2,9 +2,14 @@ import {APIResponse} from "../../../types";
 import {ValidateTokenRequestBodyType} from "@portalseguranca/api-types/account/input";
 import {FORCE_HEADER} from "../../../utils/constants";
 import {RequestError} from "@portalseguranca/api-types";
-import {AccountInfoResponse, ValidateTokenResponse} from "@portalseguranca/api-types/account/output";
+import {
+    AccountInfoResponse,
+    UserForcesResponse,
+    ValidateTokenResponse
+} from "@portalseguranca/api-types/account/output";
 import express from "express";
 import {getUserDetails, validateToken} from "../services";
+import {getAccountForces} from "../services";
 
 export async function validateTokenController (req: express.Request, res: APIResponse): Promise<void> {
     let {intents} = req.body as ValidateTokenRequestBodyType;
@@ -30,5 +35,18 @@ export async function getUserAccountDetailsController(req: express.Request, res:
         res.status(userDetails.status).json(<AccountInfoResponse>{message: userDetails.message, data: userDetails.data});
     } else {
         res.status(userDetails.status).json(<RequestError>{message: userDetails.message});
+    }
+}
+
+export async function getAccountForcesController(req: express.Request, res: APIResponse) {
+    let {nif} = req.params;
+
+    // Call the service
+    let serviceResult = await getAccountForces(res.locals.user!, Number(nif));
+
+    if (serviceResult.result) {
+        res.status(serviceResult.status).json(<UserForcesResponse>{message: "Operação concluída com sucesso", data: {forces: serviceResult.data}});
+    } else {
+        res.status(serviceResult.status).json(<RequestError>{message: serviceResult.message});
     }
 }
